Addressed user auth problem for html docs in multiple classes This is a portion of issue #340
Previously, if a url was "in" multiple classes when a user requested
access to its notes, NB would deny access entirely if there was any
such class the user couldn't access.  Now, instead, NB provides access
to notes from all classes the user can access.

This is still an incomplete fix, as there is no interface at the
client for the user to specify *which* class.  But now NB will work if
the user has access through *one* class, even if there are other
classes where the user does *not* have access.